02_airflow/requirements.txt
beautifulsoup4==4.12.3
openai==1.17.1
pandas==2.0.3
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.2
